+++ src/commands/index.ts
@@ -143,6 +143,9 @@ import { verticalGuidanceSymbols } from './a32nx/verticalGuidanceSymbols';
 import { verticalSpeedPriority } from './a32nx/verticalSpeedPriority';
 import { flyPadOS } from './a32nx/flyPadOS';
 import { simulationRate } from './support/simulationRate';
+import { dfd } from './general/dfd';
+import { direct } from './a32nx/direct';
+import { autoland } from './a32nx/autoland';
 import { holds } from './a32nx/holds';
 
 const commands: CommandDefinition[] = [
@@ -289,7 +292,10 @@ const commands: CommandDefinition[] = [
     verticalSpeedPriority,
     flyPadOS,
     simulationRate,
-    holds,
+    dfd,
+    direct,
+    autoland,
+    holds,    
 ];
 
 const commandsObject: { [k: string]: CommandDefinition } = {};